refactor(application): extract addApplicationService
