+++ src/commands/add/index.ts
@@ -1,72 +1,13 @@
-import { randomUUID } from 'node:crypto';
 import type { AddOptions, AddResult } from '../../utils/types/index.js';
 import { createAppContext, type AppContext } from '../../runtime/appContext.js';
 
 /**
  * Stores a single job submitted by an AI orchestrator (MCP-only entry point).
- *
- * Designed for AI-driven flows where the user shares a job with Claude/OpenClaw
- * (screenshot, pasted JD, or URL content). The AI caller is responsible for
- * extracting structured fields before calling this function — wolf only stores.
- *
- * Typical chain: wolf_add → wolf_score({ single: true }) → wolf_tailor
- *
- * @param options - Structured job data extracted by the AI caller.
- * @param ctx - AppContext for dependency injection; defaults to real SQLite-backed context.
- * @returns jobId for chaining into wolf_score or wolf_tailor.
+ * Thin CLI/MCP wrapper — actual logic lives in AddApplicationService.
  */
 export async function add(
   options: AddOptions,
   ctx: AppContext = createAppContext(),
 ): Promise<AddResult> {
-  const now = new Date().toISOString();
-
-  // Reuse existing company record if one exists for this name; otherwise create a new one.
-  let company = await ctx.companyRepository.getByName(options.company);
-  if (!company) {
-    company = {
-      id: randomUUID(),
-      name: options.company,
-      domain: null,
-      linkedinUrl: null,
-      size: null,
-      industry: null,
-      headquartersLocation: null,
-      notes: null,
-      createdAt: now,
-      updatedAt: now,
-    };
-    await ctx.companyRepository.upsert(company);
-  }
-
-  const jobId = randomUUID();
-  await ctx.jobRepository.save({
-    id: jobId,
-    title: options.title,
-    companyId: company.id,
-    url: options.url ?? '',
-    source: 'Other',
-    location: '',
-    remote: false,
-    salary: null,
-    workAuthorizationRequired: 'no sponsorship',
-    clearanceRequired: false,
-    score: null,
-    scoreJustification: null,
-    status: 'new',
-    error: null,
-    appliedProfileId: null,
-    tailoredResumePdfPath: null,
-    coverLetterHtmlPath: null,
-    coverLetterPdfPath: null,
-    screenshotPath: null,
-    outreachDraftPath: null,
-    createdAt: now,
-    updatedAt: now,
-  });
-
-  // JD text lives on disk (grep-friendly prose), not in SQLite.
-  await ctx.jobRepository.writeJdText(jobId, options.jdText);
-
-  return { jobId };
+  return ctx.addApp.add(options);
 }
